Experiment with "owning memo"
U cannot get this to work. It's close but I think it requires lifetime bounds on the HRTB lifetimes e.g. `F: for<'a, 'store: 'a> ...`, which currently isn't possible (seems to have been possible at one point? https://github.com/rust-lang/rust/issues/50555 Could also be just that the syntax didn't result in an error, but had no semantic meaning attached to it.)